Added functionality for converting tet meshes to spheres.  Previously, only hex meshes could be converted.
Added a new example called elastic_disk_impact.
The game plan is to replace the current series
of disk impact/failure examples with a single,
improved, example highlighting some of the more
recently added features.

A slightly longer-term game plan is to use the
example problems as a performance test suite to
be run nightly on the test machines.  The new
elastic_disk_impact example will also serve as
a driver for refactoring contact.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1302 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 2b5c87a4b2dea146ad0bcdd87bbbed5f7b09a42e