Added functionality for specifying an influence function.
The influnce function is specified via a string
passed in through the Discretization section of
the input deck, as follows:

<ParameterList name="Discretization">
  ...
  <Parameter name="Influence Function" type="string" value="One" />
  ...
</ParameterList>

Currently, the only valid option is One, which just sets
the influence function to 1.0.  The game plan is to add a
small set of built-in options, and then later to enable
arbitrary user-defined influence functions.

From the material modeler's point of view, the value of
the influence function for a given point and a given
neighbor is obtained by calling the function:

double scalarInfluenceFunction(double zeta, double horizon)

which is in material_utilities.cxx.  This function
is intended to be called from within the c-style material
evaluation functions.

It would also be possible for a material model to compute
influence function values only once, at initialization, and
store them as bond variables.  This decision is left to the
discretion of the material model developer.  All current
models call scalarInfluenceFunction() directly within each
internal force evaluation.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1307 d9297710-12f9-482b-b11f-fb677dc88bbb