Implemented user-defined influence functions.
A user defined influence function can be set
using the Influence Function parameter in the
Discretization section of the input deck.  The
allowable variables for a user-defined influence
function are zeta and horizon.

Example:

<ParameterList name="Discretization">
  ...
  <Parameter name="Influence Function" type="string" value="(horizon - zeta)/horizon" />
  ...
</ParameterList>

There are currently no tests in place for
this functionality.  Also, the error messages
need work.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1310 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 3de258e23c618d82a7c0c741d1bd94a974a84dbb