Expanded utPeridigm_SearchTree to include performance tests for the JAM and Zoltan search trees.
When run with no arguments, only the standard
unit test are run.

When run with the argument -zoltan, the zoltan
performance tests are run.

When run with the argument -jam, the JAM
performance tests are run.

The intension is to run the performance tests
under a profile build to locate hot spots, etc.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1316 d9297710-12f9-482b-b11f-fb677dc88bbb