Made a handful of modifications so that Peridigm is able to compile with the -std=c++0x option.
The main problem is that the code

using namespace std;

appears in several Trilinos header files, causing
the std namespace to be forced upon all downstream
code within Peridigm.  This then causes conflicts
between std::shared_ptr and std::tr1::shared_ptr.
This issue only appears with compilers that are
compliant with the newer C++ standards in which
shared_ptr is defined in the std namespace.  (Bug
reports have been filed with Trilinos.)



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1324 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 52e734f87287a528362f17153cefcad737507eb5