Minor refactor of surface correction factor.
Function for computing the correction factor now
explicitly resets the point at the center of the
neighborhood prior to putting the neighborhood into
pure shear.

Also, minor changes to the call to
computeSurfaceCorrectionFactor in the material models.
The current coordinates sent in are now from a temporary
vector, which eliminates any possibility of corrupting
downstream calculations by perturbing these values.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1332 d9297710-12f9-482b-b11f-fb677dc88bbb