Implemented user-specified temperature field.
The temperature field is handled as a boundary
condition.  Within the Boundary Conditions
parameter list in the input deck, the temperature
field can be specified as a function of position
and time (x, y, z, t).  It is currenlty implemented
to apply to the entire model, as opposed to only
specific blocks or node sets.

The temperature field is evaluated at every time
step and the change in temperature is fed through
to material models that register the "Temperature
Change" field.

Initial sanity check using thermoelastic material
looks good.  More testing to follow.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1355 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 12c17f92e8f956236d8d584c812ffd9fb19e2fbf