This checkin properly incorporates the 'scf' into the elastic model although it is commented out because of the large number of fails which we can easily fix;
I am still thinking about the 'scf' and how to properly use and compute it.  I
haven't fully studied the use of it just yet on basic problems that I am
looking at but initial indications are that its just giving a different nearby
solution as opposed to the 'right' solution -- this is puzzling and is leading
me to think that I am missing some basic concept; So I'm still looking at this.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1358 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 63df36b6407ead8f1d0c32cddd90f741bf89a7d5