Hooked up Rectangular_Plane bond filter.
The Rectangular_Plane filter creates a
PdBondFilter::FinitePlaneFilter under the
hood.

Bond filters are created with a Bond Filters
parameter list in the Discretization section
of the input deck.  Currently, only a single
instance of Rectangular_Plane filter may be
created.  Future work will allow for multiple
filters to be applied at once.

Added the PrecrackedPlate test, which is a
statics problem involving a plate loaded in
tension.  The plate has a precrack defined
using a Rectangular_Plane filter.  The crack
can be visualized by displaying Number_Of_Neighbors
or by visualizing the displacement field.

Bond filters are currently enabled for Exodus
meshes and Text meshes, but not for PdQuickGrid
meshes.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1360 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: c65762bd9b01a5fe307dc10d4557c7976c58e8a3