= Checkin Feb 22, 2012 == Checkin on 'SCF' formerly known as the 'DSF'
This is an improvement; the simple 7x7x7 cube in simple shear now gets
exactly the correct strain energy density on every point;  This is only true if
I artificially turn off the dilatation.  Even when we have a state of pure shear 
the dilatation still comes out non-zero on a coarse mesh and this is somewhat of a 
problem.  But I was able to confirm that the scf calculation is doing what I 
was hoping.

There seems to be a bug in the probe or there is something else going; currently 
I had to shut off the probe in xz, yz for the simple shear case.  Need to look into
that further.

BUMMER, when I use this on the 'tensile_test' there is basically no correction.
I have another idea for that but I wanted to get this checkin.

Also, I set the default value of m_applySurfaceCorrectionFactor to 'false' otherwise there 
would be a lot of failures.  When I get all of this sorted out then we will have 
to visit all the tests.

I also had to de-activate a couple of my tests related to the quadrature and scf calculation.
Will look at those later.






git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1364 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: a524b8058c8840f671f29b98a98ad22c1a3dbab9