Progress on refactoring rebalance and contact.
With this checkin, a refactored proximity search
is now applied when reading Exodus mesh files.
The difference from the user's perspective is that
the decomposition supplied by the user is maintained.
Previously, a new RCB decomposition was performed
at initialization.  This is no longer the case,
instead the output discretization (*.e) matches
the input discretization (*.g)

This does not yet effect contact.  In that case, you'll
still get repeated rebalances.  This checkin is a step
toward that but there's still quite a way to go.

QuickGrid and text file discretizations are not affected
by these changes.

This checkin contains both the old and new versions of
the initial proximity search for Exodus discretizations.
Switching between them is controlled by a #define.  I'm
checking it in this way so I'll have the old version
archived.  I'll rip it out and clean up the code prior
to the next checkin.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1407 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 46c8edcebe3cacb9ac3fcdc8ccdfb43b05eff752