Zero out tangent values for kinematic BCs by row access instead of column
For quasi-static analysis, zeroing out the rows and columns
for kinematic BCs accounts for roughly 90% of the execution time
(for problems with large numbers of BC nodes). This commit rearranges
how the zeros are inserted to be row oriented rather than column.
The result is a 4x+ speedup in the execution time for the test problems
I was using.

Notes:

A category for "Applying BCs" should be added to the timing output to keep 
track of these costs (at least for implicit or quasi-static).

The general way kinematic BCs are handled could use some attention. Even with
these changes it still takes up roughly 70% of the execution time. All this
expense is coming from reinserting values into the jacobian. There's a few 
alternative approaches that could be used instead.
 - avoiding assembly for BC nodes in the first place (why compute what 
   you eventually throw away?)
 - excluding BC rows/columns from the linear system (reduces problem size 
   and can lead to better conditioning, no need for normalized diagonal entries)
 - other ideas?




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1412 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 2ace1ee1b172495622f620a3f8007a6b627c7e6e