Example using the new input deck format.
This is a copy of the "tensile_test" example, but
using the new Peridigm input deck format.  You will
now see a file "tensile_test.input" that includes
the new input deck commands. It demonstrates comments
and Aprepro functionality.  The file extenstion can be
anything except .xml.  If there is a .xml file is
given the old input parser will be used.


git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1420 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: d51aff16322d1ecd68df91501857e71f9664d4f6