Added XML ripper script.
This Python script will aid in converting old XML
input decks to the new format.  It is not a 100
percent solution though.  It mainly just strips off
the XML markup, the user will still need to manually
remove the quotations from around numbers that are
intended to be doubles or integers.  Also, some
of the indention may need to be adjusted after running
the script.


git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1421 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: b16a44cb3a95c46499cd5c69685b013b07c3d1a8