Fixed build with votd Trilinos.
The issue was the std namespace, which bit
me once before.  Several packages, including
one that we use (NOX?) had a "using namespace
std;" in a header file, and this caused the
std namespace to be used throughout all of
Peridigm, whether we wanted it or not (in
general we do not, because it causes issues
with C11 and shared_ptr, and is not good
practice in general to put a using statement
in a header file).  The using
namespace statement was recently removed
from the offending Trilinos header file but
is obviously still there in older releases.

I had a vector declared without the std:: on
front, and it complied with older versions of
Trilinos but failed with VOTD (as it should).



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1446 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: a26dab06148ea82f8f0f5402fd17138d41a79da5