Refactored code to allow for block-by-block assignment of the horizon.
**** IMPORTANT ***

This commit will break all existing input decks.  The Horizon
parameter has been moved from the Discretization section to
the Blocks section.  A Horizon parameter must now be specified
for each block.  The parameter entry itself is unchanged, so you
can just cut-and-paste it from the Discretization section into
each Block definition.

******************

Note that the horizon is a parameter for numerous parts of the
code:  neighborhood search, material model, contact model,
and possibly the damage model.  The code has been refactored
such that the block's horizon value is passed down to everything
associated with the block (e.g., the material model).

One motivation for this work was to enable a simulation in which
two blocks with different horizons interact via contact.  This
should work now.  Be aware that the contact models associated with
the two blocks will have different horizons, which affects the
magnitude of the contact force associated with each block.

Also note that if a simulation has two blocks with different horizons,
and the blocks are bonded to each other, there will asymmetric bonds
(i.e., bonds from A->B for which there are no corresponding bonds from B->A).
Use this approach at your own risk!

This refactor brought to light some architectural shortcomings that
create order-of-operations issues.  There are workarounds in place
for these, but more refactor work is required to get them right.  For
example, the Discretization objects are currently responsible for
generating the neighborhood lists, and this should be moved outside
the Discretizations to allow for better handling of multiple horizons.
This situation has led to ugly code in both Peridigm and the
Discretization objects themselves.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1451 d9297710-12f9-482b-b11f-fb677dc88bbb