Added a verification test for simulations with multiple horizons.
The primary goal of the test is to check the
assignment of a horizon to each block and to verify
the resulting number of neighbors for each element.

The test also exercises the Text File discretization
and the ability to provide user-specified time steps
for quasi-static analyses.

Fixed a bug in the Text File discretization
which caused problems when using multiple blocks in
parallel.

Added a check for NaNs in the quasi-static residual
calculator.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1457 d9297710-12f9-482b-b11f-fb677dc88bbb