Progress on contact refactor.
Moved contact functionality out of Peridigm.cpp
and into a newly-created ContactManager.  There
should be no change in overall functionality or
analysis results.

This checkin contains both the old code and the
new code.  There is a #define CONTACT_REFACTOR
that toggles between the old and new code.  This
commit will be quickly followed by a cleaned-up
version with only the new code.  I wanted to
have a version of the code in the repository with
both versions for testing purposes.  The ContactManager
is currently a mess.  It will be cleaned up in the
near future.

All the tests pass and the elastic_disk_impact
example problem gives the same results with both
the old and new code.  If anyone sees changes in
results for other analyses, let me know.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1462 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 2a61185d492879cc7377a70f061916256b3aaf74