Initial implementation of an elastic correspondence material model.
Some refactoring is still required, and suppression
of zero-energy modes needs to be added.

Added a test with prescribed uniaxial homogeneous strain.
All displacement degrees of freedom are fully prescribed.
The test checks the force evaluation and recovery of the
elastic modulus.  Fully prescribing the displacement
circumvents the zero-energy-mode issue.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1472 d9297710-12f9-482b-b11f-fb677dc88bbb