Disabled preconditioner for quasistatics.
The preconditioner scheme in Peridigm started
exhibiting problems with Trilinos 11.2.5 (Jul-11-2013).
At this time it is not clear what the root cause of the
difficulty is.  For the record, the preconditioner
scheme in Peridigm appears to function properly with
Trilinos 11.2.4 (Jun-20-2013).

The preconditioners will be re-enabled once
we get the ifpack issue straightened out.

If you're running with an older version of
Trilinos (<= 11.2.4) and want to enable the
precoditioners, just change

bool usePreconditioner = false;

to

bool usePreconditioner = true;

in Peridigm.cpp.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1485 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 4274449720d2e50cdcf486a89a5637d54de8e53f