Added basic memory stats to program output
The usual suspects for memory use (allocating the
jacobian, creating the neighbor list, etc.) are now
tracked for heap usage and statistics are provided
at the end of a simulation. For more detailed 
profiling, the user should use a tool like Valgrind.

To add stats for a specific location in the code, add
the include #include "Peridigm_Memstat.hpp" and 
the following lines of code:

PeridigmNS::Memstat * memstat = PeridigmNS::Memstat::Instance();
memstat->addStat("<description goes here>");

This will add a stat with your description to the output.
If your stat is executed multiple times, only the max 
use is reported.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1487 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: a31c26d6ccb76c895d7a8becc3014c72a6d22105