Created base class for correspondence material models.
Correspondence models should now derive from the
CorrespondenceMaterial base class.  The derived class
must implement the function

void computeCauchyStress(const double dt,
                         const int numOwnedPoints,
                         const int* ownedIDs,
                         const int* neighborhoodList,
                         PeridigmNS::DataManager& dataManager);

The general idea is that a correspondence material
model that is derived from CorrespondenceMaterial
will compute the Cauchy stress based on:

1) The deformation gradient at step N.
2) The deformation gradient at step N+1.
3) The time step.
4) Whatever state variables are managed by the derived class.

The calculation of the deformation gradient, etc., is
handled by the CorrespondenceMaterial base class, so all
the derived classed need to do is implement computeCauchyStress().

The ElasticCorrespondenceMaterial has been refactored to
follow this pattern.  It can be used as a template for
implementing other correspondence material models.

Bond failure is not currently implemented for correspondence
material models.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1498 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 760bd54b8e14eccf4b49894ddb962b1ab6ab88db