Refactor Correspondence Material Interface.
This is a minimally tested refactor of the Correspondence Material Interface.
It uses the Flanagan and Taylor (1987) algorithm to perform a left polar
decomposition and solve for a rotation tensor that can then be used to
"unrotate" the rate-of-deformation tensor.  This is then supplied to the
constitutive model to update the corotational Cauchy stress.

The material model developer now only needs to implement an incremental stress
update algorithm that takes a rate-of-deformation tensor as a measure of
of strain-rate and returns an updated Cauchy stress.  All of the kinematics
are handled uniformly across correspondence material models.

The Taylor and Flanagan algorithm was implemented as described in the 1987
paper using the "explicit time integration" formulation.  It may be
necassary to implement the "implicit" version for QS and implicit solves to
be performed in a stable way.  This is only a slight change to the algorithm,
namely replacing the solve of Eq. (39) in the paper with Eq. (44)

Squashed commit of the following:

commit 0739afc484d5a9b0e8741872a0da783ef8347fd7
Author: John Foster <john.foster@utsa.edu>
Date:   Wed Sep 11 17:28:28 2013 -0500

    Temporarily Disabled ElasticCorrespondenceMaterial tests

commit 474512bc7b1a4b443801a4aeb1d41671d0d6a142
Author: John Foster <john.foster@utsa.edu>
Date:   Wed Sep 11 17:06:53 2013 -0500

    Fix more build errors

commit d80580d45d7f209ce05a542ea98e48cb185f5fbc
Author: John Foster <john.foster@utsa.edu>
Date:   Wed Sep 11 16:16:52 2013 -0500

    Fix build

commit 5f4e361609361842ed7bfe0fb6ca136d7652f1e9
Author: John Foster <john.foster@utsa.edu>
Date:   Wed Sep 11 15:56:54 2013 -0500

    Fix build

commit 5f7ff76797df64800c79b30f8523c09fa39e342d
Author: John Foster <john.foster@utsa.edu>
Date:   Wed Sep 11 13:55:27 2013 -0500

    Code cleanup

commit 5e6e82be143501abce76dbb30fa1577efeb3a8b2
Author: John Foster <john.foster@utsa.edu>
Date:   Wed Sep 11 13:28:40 2013 -0500

    Add cxx file for new elastic correspondence model

commit ae3ed69dcc01699e46d4498eb3d5b2cceea1e3c0
Author: John Foster <john.foster@utsa.edu>
Date:   Wed Sep 11 11:05:23 2013 -0500

    Add methods for rotating and unrotating Cauchy Stress.

commit ffe033239bb94380a9eb63da5841887b413bdcd7
Author: John Foster <john.foster@utsa.edu>
Date:   Tue Sep 10 22:59:29 2013 -0500

    Add Rotate and UnrotateTensor methods to correspondence.cxx

git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1507 d9297710-12f9-482b-b11f-fb677dc88bbb