Adds 'nearest neighbor search' to kdtree.
Test is also added.  Test consists of a tree with a million points.  
For each point 'p' in the tree, another point 'q' is created which 
is randomly positioned nearby so that 'p' is the nearest neighbor 
of 'q'.  Hence the answer for each nearest neighbor search 
is known apriori.  



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1519 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: b32183f90b456c8154f8c32174ce8f00467bd1b7