Reactivated ElasticCorrespondenceQSTension test.
Currently, the output of the hourglass force
vector is broken in parallel.  I believe this
error affects only output; the underlying algorithm
is functioning correctly.  The issue is that the
hourglass force is not summed across processors
in parallel.  On processor, it is added to the
(standard) force vector, which is subsequently
assembled across processors, so all is well in
terms of the internal force calculation.  But
the hourglass forces themselves are not assembled,
hence the output error running in parallel.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1531 d9297710-12f9-482b-b11f-fb677dc88bbb