Fixed bug in NOX-based QS solver and reactivated the (one and only) regression test of the NOX-based solver.
In the NOX-based solver, the vector containing the
displacement increment for a given load step has
zeros for the dof corresponding to kinematic
boundary conditions.  This is by design, perhaps
out of necessity for the black-box nature of using
NOX to solve the system.  A consequence of this is
that one cannot back out the nodal velocities from
the displacement increment, because the velocities
for dof with kinematic bc will be incorrectly
calculated as zero.  The fix I implemented simpily
stores the velocities for these dof separately, and
then sums them into the main velocity vector. Having
the correct velocities is required for intenal force
calculations using the correspondence models, and
possibly for the viscoelastic material as well.
There are currently no tests of the NOX-based solver
with these constitutive models.

The line search originally set up to improve the
predictor in NOX has been removed.  It will not
work with the above-mentioned velocity issue.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1533 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 9bdc758efca61e805c2c48c0e8f924f6c114aa3a