pre_compute() method for compute classes
Added a method, pre_compute() that can be called
by the compute manager after initialization, but prior
to compute(). This method is executed once by the compute manager,
not at every step like compute(). It can be used to
do initialization calculations at the first step that 
do not need to be repeated at each step.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1545 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 9e7fae28a1688cf2974edc469870f9a4d6905819