Hopefully fixed the build issue on Latrobe
There was a capitalization issue with a header file included in the file 
below. My office-mate Brent P. informs me that since Macs have a case-insensitive
file system, this error would not show up until it was built on a linux
machine like latrobe. We'll see in a few minutes if the fix works...



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1549 d9297710-12f9-482b-b11f-fb677dc88bbb