First attempt at getting bond breaking working in Implicit/QS
Seems that most of the issues with getting a converged result
for bond breaking in implicit/QS are related to zero pivots, etc
associated with a node having too few bonds to be stable. As bond
breaking progresses through the mesh for a given step, instances
occur for which all the bonds of a node break (or all but two).
The solver then fails. 

In this first pass, any "rank deficient" nodes are taken out of 
the linear system by pushing them into a boundary condition 
node set: RANK_DEFICIENT_NODES. All the bonds are broken for
these nodes so it effectively removes them from the analysis.

A new damage model is used to do the tasks above: Interface Aware.
This model does not allow bonds to break between blocks if one 
of the blocks does not have a damage model. This damage model
also does the checking of nodes if they meet the rank deficient
criteria.

A flag was also introduced that enables the user to turn off
solver heuristics. For the examples tested with this commit,
none of the solver heuristics were needed to get convergence.
Examples have been tested that have a crack propagating through
the entire domain in a single step and the solver manages to
converge (if the Interface Aware damage model is used).
To activate this flag, set "Disable Heuristics" to true in the 
solver parameters.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1556 d9297710-12f9-482b-b11f-fb677dc88bbb