Changed timing benchmarks for disk_impact performance test.  The target is now set at 1350 seconds and the tolerance was raised to 7%.
Last night's tests had the disk_impact problem
running in 1300 seconds with Trilinos 10.10.1
and 1400 seconds with VOTD Trilinos, a difference
of over 7%.  It is really impossible to draw
conclusions based on a single data point, but this
does highlight the difficulty in performance
testing in general.

I will start thinking about ways to track
performance over time.  A plot of runtime
that stretches over several weeks might be
useful.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1568 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: c7251e1a003cadad176bf3d5e4f69af95bdb02a7