Refactored the short-range force contact model to reduce the number of calls to sqrt().
Vtune profiling of the disk_impact example problem
showed that sqrt() was a very dominant portion of the
run time.  This change speeds up the disk_impact
example problem by roughly 20%.  Speed-up in
general will be problem dependent.  Problems
in which the contact search radius is larger than
the contact radius will show good speed-up.  If the
search radius and the contact radius are equal, there
will be no speed-up.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1570 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 0dd6c6d6133a5a882bbe853fe1175167ad675396