Switched a stl container type in the contact algorithm based on vtune profiling.
The profiler identified a bottleneck in the
contact algorithm in which a neighborhood list
is searched to determine if a given element id
is in that neighborhood.  This was originally
done using a std::list and has been switched to
a std::set.  The performance gains are very
significant for this particular sort (typical
neighbor lists have roughly 400 entries).

One lesson learned:  Based on a bake-off of
various stl containers, etc., I found that for
arrays of this length the member version of find
(i.e., std::set::find()), is massively faster
than the non-member function (e.g., std::find()).
The fact that the member function is faster is
not surprising, but the huge difference in performance
was an eye-opener.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1573 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 22c7e9bea8da817c9a6637a305a70c0574a1c1e5