Fixed fragmenting_cylinder example problem.
The initial velocities were incorrectly applied
to All Sets, where as they should have been applied
to Full Domain.

Also, replaced the user defined time step with
a safety factor that is applied to the automatically
calculated critical time step.  This resulted in a
massive speed-up, as the user-defined time step was
very conservative.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1585 d9297710-12f9-482b-b11f-fb677dc88bbb