Implemented a HorizonManager.
The purpose of the HorizonManager is to consolidate
the horizon-parsing code that was previously sprinkled
all over the place into one class and to extend Peridigm's
ability to handle various horizon options.

The HorizonManager includes the ability to
parse an expression for the horizon.  This is needed
for ongoing Sandia internal research.

Also, added a horizon mothership vector that stores
the horizon for each point in the discretization.
Previously, a single value of the horizon was assigned
to each block.  This has been generalized to allow
the horizon to be defined at each point.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1600 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 130483e9e348f29d8df81dd30ea976e717fd342e