Added some scripts for querying the performance test results and plotting accumulated performance data over time.
The script

query_performance_test_results.py

is meant to be run in the directory where
the tests are run.  It finds the log files for
the performance tests and prints the results
in a table to stdout.

These tables will be appended to a file on latrobe
that contains a running tally of performance test data.

The script

plot_performance_test_results.py

is a (pretty crude) utility for generating plots
with gnuplot.  Perhaps this will eventually be
replaced with a script to generate plots that can
appear automatically on the peridigm web site.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1614 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 0bdd31cb57bb16493346c20abbf22e55f775deb7