Removed the vast majority of paths from #include statements and added the corresponding directories to an include_directories() command in the main CMakeLists.txt file.
This improves our ability to install libPeridigmLib.a and
call it from another code.  In this case, the cmake install
command puts all the Peridigm headers into a single directory,
as opposed to trying to reconstruct the directory hierarchy.
Removing paths from #include statements allows the build system
to function in the case of either a flat header directory or the
full hierarchy.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1635 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 0d0589d2847d0686debcec9d96983892ec89e203