Kokkos version of elastic material compute force
This is the initial implementation of a kokkos functor
for the force calculation for an elastic material. A compiler flag,
USE_KOKKOS, was added to conditionally add/remove the kokkos
version of the force calculations.

There has been no effort in this implementation to get the manycore
parameters optimized (thread count, teams, etc.). Once we have identified 
what machine the GPU optimization work will be performed on, we
can revisit this.

Also note that if you have USE_KOKKOS enabled, it won't play nice with
MPI parallel runs of the elastic force calculations. More work will need
to be done to manage this conflict.

Lastly, since there isn't a mechanism in kokkos yet for unmanaged memory
on the host being shared through the kokkos interface to device memory, the
data from all of the peridigm fields has to be manually deep copied. The kokkos
developers are working on addressing this so in the future we can just
construct a view of the existing data without a manual copy.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1667 d9297710-12f9-482b-b11f-fb677dc88bbb