Fixed Dakota test.
There were several things broken in this test, I don't
think it has run properly for some time.

In this test, Dakota is used to solve for the bulk modulus
in the sense of using an optimization routine to calibrate
a constitutive model.

The initial guess for the bulk modulus was set to the truth
value, meaning that Dakota had nothing to do.  It just
evaluated the objective function, got a value of zero, and
exited.  This doesn't make much sense.  The initial guess is
now set to 100.0e9, which differs from the truth value of 130.0e9.

This test contains several scripts, one of which calls exotxt
followed by a series of shell commands to extract nodal forces from
the Peridigm output file (it might be better to use something like
blot or a exodus python script that doesn't rely on the exact positions
of values in the exotxt output text file).  The recent addition of QA
data to the Peridigm output broke this script, because the output
of exotxt now has four additional lines (the QA records).  For some
reason, both the line number and the column values were incorrect
in the script that parses the exotxt output.  The addition of QA
records explains the incorrect line numbers, but not the incorrect
column values.

Finally, the tolerance for comparing Dakota's solution for the bulk
modulus against the truth value was too tight.  It has been relaxed
to a more reasonable value.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1671 d9297710-12f9-482b-b11f-fb677dc88bbb