****IMPORTANT****
SIMULATION RESULTS MAY CHANGE.

This commit switches the hex volume and centroid
calculations from the old approach to the new
approach.  This will affect simulations that read
nonuniform hex meshes from Exodus files.  Conversion
of a non-rectangular hex element to a sphere element
will give different results for the centroid.  If the
hex contains nonplaner faces, the volume will also be
different.  In most cases, the differences will be
small, but it is likely that they will change
results for nontrivial simulations.

The new centroid and volume calculations operate by dividing
the hex into 24 tets.  This is accomplished using the 8 hex
nodes, plus 6 nodes located at the barycentric centers of the
faces, plus a node in the barycentric center of the hex.  This
approach removes orientation dependence (which is present if
you divide the hex into six tets).  If the hex has planar
faces, the volume calculation is exact and unambiguous.  If
there is one or more nonplanar face, the nonplanar faces are
represented with 4 triangles.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1679 d9297710-12f9-482b-b11f-fb677dc88bbb