Infrastructure for internal interfaces between elements
In the Discretization section of the input, the user can 
specify "Construct Interfaces" as true to build a mesh
of quads between hexes, and tris between tets. 

The mesh is output to a file called Interfaces.e. In a future
commit we can adapt the filename to be specified in the input.

The data for these interfaces is stored in an
InterfaceData class, which currently holds the right and
left neighboring elements, the connectivity of the 
interface element, and field data. There is only one
field for now, but as more fields are added this may
need to be reworked for efficiency and memory use. Interfaces
are only constructed between elements that share a face
in the original exodus mesh. 

Nothing is currently computed on the interface for this
commit. It is just the basic implementation of the
InterfaceData class.

A new test was also added to ensure that the interfaces
are constructed properly for meshes with hexes or tets or
a combination of the two in serial and parallel.




git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1682 d9297710-12f9-482b-b11f-fb677dc88bbb