Added a compute class for bond visualization.
The class writes a vtk file (one for each processor)
that contains the bonds in the reference configuration.
The bonds are displayed as VTK_LINE elements.  The idea
is that you can load this file into Paraview in addition
to the standard exodus file generated by Peridigm.

Intended mainly for debugging and PowerPoint.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1707 d9297710-12f9-482b-b11f-fb677dc88bbb