Altered function parsing to support bond fide C code.
Function strings may be used in the Peridigm input deck
for boundary conditions, influence functions, and horizon
functions.  The internal mechanism for processing them
used to be muParser, and was recently changed to rtcompiler.

Under the old system, the user input a string that represents
the right-hand side of an equation.  For example, if you wanted
your boundary condition function to be a quadratic function of
x, you could use

2.0*x*x

as your string.  This is what muParser expects.

The rtcompiler expects the entire equation, not just the
right-hand side.  For the rtcomplier, a valid expression is

value = 2.0*x*x

What Peridigm does now is search the expression for the
term 'value'.  If this term is not present, the string
'value = ' is inserted at the beginning of the expression
prior to passing it to the rtcompiler.  This provides
backwards compatibility and avoid issues with the horizon
manager (I'll spare you the details).

So, if your expression is a simple one, you can do what
you always did, and Peridigm will handle it properly internally.

If your expression is more complex, for example the boundary
conditions int the twist_and_pull example problem, you'll need
to replace the expression with bona fide C code that evaluates
the variable 'value'.

The good news is that the rtcompiler syntax is much more general
and will allow users to input extremely complex expressions if they
are so inclined.  The bad news is that this change may break old
input decks that have nontrivial function expressions.

The twist_and_pull example is currently broken.  Fixing this is
the next order of business.



git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1717 d9297710-12f9-482b-b11f-fb677dc88bbb

Former-commit-id: 02dc1bb21b17e6ed406bf084442541f3723ab368