Updated timing values in .perf files so that the performance tests will pass again.
All the tolerances were increased for the time being.  This will
allow a few day's worth of timing results to be collected.  Then
the tolerances can (hopefully) be tightened up again.

The run time for the twist_and_pull test increased by a factor of
two as a result of switching to rtcompiler.  The impact was much
smaller for the other tests.

Also, switching to rtcompiler resulted in a different version of rand()
being called by the function parser when processing the input
deck for the fragmenting_cylinder test.  Here, rand() is used to peturb
the initial velocities of the nodes throughout the model.  Rtcompiler
calls the standard C library version of rand(), and does so in a way
that is not repeatable.  Thus the results of the fragmenting_cylinder
test are different every time you run the simulation.  With muparser,
we had rigged the rand() function to be repeatable.  Something similar
needs to be done in Peridigm for rtcompiler.

Currently, the disk_impact, tensile_test, and twist_and_pull performance
tests check timing data and check the exodus results file against a gold
file that is stored on the test machine.  The fragmenting_cylinder test
checks only timing data.